device-handle-bot/requirements.txt
python-dotenv
confluent-kafka
python-telegram-bot==21.6
